Reimplement showing of authors into its own function "show-authors".
I have come to the conclusion that i have merged showing of authors and music too soon.
I believe that while code for both will probably be repeated, some code will need to be
different. It was too soon of me to try to handle both cases in the same function.